Moves constants to main com.zgillis.nexochat package.
