+++ src/com/zgillis/nexochat/server/Client.java
@@ -1,5 +1,6 @@
 package com.zgillis.nexochat.server;
 
+import com.zgillis.nexochat.*;
 import java.io.*;
 import java.net.*;
